requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
